requirements.txt
argparse
numpy
torch
monai
torchvision
pandas
scikit-learn
tqdm
pycox
sklearn-pandas
torchtuples
wandb
umap-learn
lifelines
seaborn
matplotlib
simpleitk
PyYAML